+++ projection/moves/move_left.c
@@ -13,8 +13,8 @@
 
 void	move_left(t_mlx *wind)
 {
-	float	y_step;
-	float	x_step;
+	double	y_step;
+	double	x_step;
 	int		y_map;
 	int		x_map;
